requirements.txt
psycopg2
pillow
numpy
flask
opencv
pytorch
torchvision
facenet_pytorch